+++ Sources/RoktUXHelper/Data/Expansion/Domain/DomainMappableParent.swift
@@ -68,3 +68,6 @@ extension CatalogStackedCollectionViewModel: DomainMappableParent {}
 
 @available(iOS 15, *)
 extension CatalogResponseButtonViewModel: DomainMappableParent {}
+
+@available(iOS 15, *)
+extension CatalogDevicePayButtonViewModel: DomainMappableParent {}