+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `CatalogDevicePayButton` node supported
+- `CartItemDevicePay` UX event supported
 - Hide the Creative Response Component when the actionType is external
 
 ### Fixed